+++ tests/integration/fixtures/processing.py
@@ -63,7 +63,6 @@ def processing_config(get_topic_name):
                 "monitors": get_topic_name("monitors"),
                 "spans": get_topic_name("spans"),
                 "profiles": get_topic_name("profiles"),
-                "metrics_summaries": get_topic_name("metrics_summaries"),
                 "feedback": get_topic_name("feedback"),
             }
 
@@ -354,11 +353,6 @@ def profiles_consumer(consumer_fixture):
     yield from consumer_fixture(ProfileConsumer, "profiles")
 
 
-@pytest.fixture
-def metrics_summaries_consumer(consumer_fixture):
-    yield from consumer_fixture(MetricsSummariesConsumer, "metrics_summaries")
-
-
 class MetricsConsumer(ConsumerBase):
     def get_metric(self, timeout=None):
         message = self.poll(timeout=timeout)
@@ -521,24 +515,6 @@ class ProfileConsumer(ConsumerBase):
         assert message.error() is None
 
         return msgpack.loads(message.value()), message.headers()
-
-
-class MetricsSummariesConsumer(ConsumerBase):
-    def get_metrics_summary(self):
-        message = self.poll()
-        assert message is not None
-        assert message.error() is None
-
-        return json.loads(message.value())
-
-    def get_metrics_summaries(self, timeout=None, n=None):
-        metrics_summaries = []
-
-        for message in self.poll_many(timeout=timeout, n=n):
-            assert message.error() is None
-            metrics_summaries.append(json.loads(message.value()))
-
-        return metrics_summaries
 
 
 class CogsConsumer(ConsumerBase):